+++ android/src/main/kotlin/com/caramelpoint/mercado_pago_integration/MercadoPagoHelper.kt
@@ -1,6 +1,7 @@
 package com.caramelpoint.mercado_pago_integration
 
 import android.app.Activity
+import android.app.Activity.RESULT_CANCELED
 import android.content.Intent
 import com.google.gson.Gson
 import com.mercadopago.android.px.core.MercadoPagoCheckout
@@ -10,21 +11,15 @@ import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.PluginRegistry
 
-class MercadoPagoHelper : PluginRegistry.ActivityResultListener {
+
+class MercadoPagoHelper(activity: ActivityPluginBinding) : PluginRegistry.ActivityResultListener {
 
     private var requestCode = 7717
 
-    constructor(activity: ActivityPluginBinding) {
-        this.activityPluginBinding = activity
-        activity.addActivityResultListener(this);
-    }
-
-    private var activityPluginBinding: ActivityPluginBinding
+    private var activityPluginBinding: ActivityPluginBinding = activity
     var result: MethodChannel.Result? = null
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?): Boolean {
-        println("ACTIVITY RESULT")
-        println(requestCode.toString())
         if (requestCode == this.requestCode) {
             onMercadoPagoResult(resultCode, data)
         }
@@ -36,25 +31,24 @@ class MercadoPagoHelper : PluginRegistry.ActivityResultListener {
     }
 
     private fun onMercadoPagoResult(resultCode: Int, data: Intent?) {
-        var response = HashMap<String, Any>()
+        val response = HashMap<String, Any>()
         response["resultCode"] = "$resultCode"
-        if (isSuccessResult(resultCode)) {
-            println("SUCCESS")
-            response["payment"] = getPaymentFromData(data)
-            println( response["payment"])
-        } else if (isErrorResult(resultCode, data)) {
-            println("ERROR")
-            response["error"] = getErrorFromData(data)
-            println( response["error"])
+        when {
+            isSuccessResult(resultCode) -> {
+                response["payment"] = getPaymentFromData(data)
+            }
+            isErrorResult(resultCode, data) -> {
+                response["error"] = getErrorFromData(data).message
+            }
+            else -> {
+                response["error"] = "Canceled"
+            }
         }
-        var json = Gson().toJson(response);
-        println("JSON RESULT")
-        println(json)
-        this.result?.success(json)
+        this.result?.success(Gson().toJson(response))
     }
 
     private fun isErrorResult(resultCode: Int, data: Intent?) =
-            resultCode == Activity.RESULT_CANCELED && data != null && data.extras != null
+            resultCode == RESULT_CANCELED && data != null && data.extras != null
 
     private fun isSuccessResult(resultCode: Int) =
             resultCode == MercadoPagoCheckout.PAYMENT_RESULT_CODE
@@ -65,5 +59,9 @@ class MercadoPagoHelper : PluginRegistry.ActivityResultListener {
 
     private fun getErrorFromData(data: Intent?): MercadoPagoError {
         return data!!.getSerializableExtra(MercadoPagoCheckout.EXTRA_ERROR) as MercadoPagoError
+    }
+
+    init {
+        activity.addActivityResultListener(this)
     }
 }